+++ data/src/main/java/com/egiwon/data/di/NetworkModule.kt
@@ -1,0 +1,55 @@
+package com.egiwon.data.di
+
+import com.egiwon.data.datasource.RandomImageService
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.Interceptor
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+class NetworkModule {
+
+    @Provides
+    @Singleton
+    fun provideHttpLoggingInterceptor() = HttpLoggingInterceptor().apply {
+        level = HttpLoggingInterceptor.Level.BODY
+    }
+
+    @Provides
+    @Singleton
+    fun provideGsonConverterFactory(): GsonConverterFactory = GsonConverterFactory.create()
+
+    @Provides
+    @Singleton
+    fun provideHttpClient(
+        loggingInterceptor: HttpLoggingInterceptor,
+        interceptor: Interceptor
+    ) =
+        OkHttpClient.Builder()
+            .addInterceptor(loggingInterceptor)
+            .addInterceptor(interceptor)
+            .build()
+
+    @Provides
+    @Singleton
+    fun provideRetrofit(
+        converterFactory: GsonConverterFactory,
+        client: OkHttpClient
+    ): Retrofit = Retrofit.Builder()
+        .baseUrl("https://api.unsplash.com/")
+        .client(client)
+        .addConverterFactory(converterFactory)
+        .build()
+
+    @Provides
+    @Singleton
+    fun provideRandomImageService(retrofit: Retrofit): RandomImageService =
+        retrofit.create(RandomImageService::class.java)
+}
